Refactored some code, added but not fully implemented remaining model classes, added customers table view. Code is messy, but will clean up later.

+++ src/controller/MainController.java
@@ -12,7 +12,9 @@ import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
-import model.Appointments;
+import model.Appointment;
+import model.Customer;
+import utils.ErrMsg;
 
 import java.io.IOException;
 import java.net.URL;
@@ -25,79 +27,134 @@ import java.util.ResourceBundle;
  */
 public class MainController implements Initializable {
 	/**
-	 * The Appointments table.
+	 * The Appointment table.
 	 */
 	@FXML
-	private TableView<Appointments> appointmentsTable;
+	private TableView<Appointment> appointmentsTable;
 	/**
 	 * The Appointment id column.
 	 */
 	@FXML
-	private TableColumn<Appointments, Integer> appointmentIDColumn;
+	private TableColumn<Appointment, Integer> appointmentIDColumn;
 	/**
 	 * The Title column.
 	 */
 	@FXML
-	private TableColumn<Appointments, String> titleColumn;
+	private TableColumn<Appointment, String> titleColumn;
 	/**
 	 * The Description column.
 	 */
 	@FXML
-	private TableColumn<Appointments, String> descriptionColumn;
+	private TableColumn<Appointment, String> descriptionColumn;
 	/**
 	 * The Location column.
 	 */
 	@FXML
-	private TableColumn<Appointments, String> locationColumn;
+	private TableColumn<Appointment, String> locationColumn;
 	/**
 	 * The Contact column.
 	 */
 	@FXML
-	private TableColumn<Appointments, String> contactColumn;
+	private TableColumn<Appointment, Integer> contactColumn;
 	/**
 	 * The Type column.
 	 */
 	@FXML
-	private TableColumn<Appointments, String> typeColumn;
+	private TableColumn<Appointment, String> typeColumn;
 	/**
 	 * The Start time column.
 	 */
 	@FXML
-	private TableColumn<Appointments, LocalDate> startTimeColumn;
+	private TableColumn<Appointment, LocalDate> startTimeColumn;
 	/**
 	 * The End time column.
 	 */
 	@FXML
-	private TableColumn<Appointments, LocalDate> endTimeColumn;
+	private TableColumn<Appointment, LocalDate> endTimeColumn;
 	/**
 	 * The Customer id column.
 	 */
 	@FXML
-	private TableColumn<Appointments, Integer> customerIDColumn;
+	private TableColumn<Appointment, Integer> customerIDColumn;
 	/**
 	 * The User id column.
 	 */
 	@FXML
-	private TableColumn<Appointments, Integer> userIDColumn;
+	private TableColumn<Appointment, Integer> userIDColumn;
+
+	private static Appointment selectedAppointment;
+
+	/**
+	 * Gets selected appointment.
+	 *
+	 * @return the selected appointment
+	 */
+	public static Appointment getSelectedAppointment() {
+		return selectedAppointment;
+	}
 
 	/**
 	 * The Add appointment button.
 	 */
-	@FXML private Button addApptBtn;
+	@FXML
+	private Button addApptBtn;
+	@FXML
+	private Button updateApptBtn;
+
+
+
+
+
+
+
+
+	/**
+	 * The Customer Table
+	 */
+	@FXML private TableView<Customer> customersTable;
+	/**
+	 * The Customer ID column
+	 */
+	@FXML private TableColumn<Customer, Integer> customerIDColumn2;
+	/**
+	 * The Customer Name column
+	 */
+	@FXML private TableColumn<Customer, String> customerNameColumn;
+	/**
+	 * The Phone Number column
+	 */
+	@FXML private TableColumn<Customer, String> phoneNumColumn;
+	/**
+	 * The Address column
+	 */
+	@FXML private TableColumn<Customer, String> addressColumn;
+	/**
+	 * The State column
+	 */
+	@FXML private TableColumn<Customer, String> stateColumn;
+	/**
+	 * The Postal Code column
+	 */
+	@FXML private TableColumn<Customer, String> postalCodeColumn;
+
+
+
+
+
+
 
 	@Override
 	public void initialize(URL url, ResourceBundle resourceBundle) {
 		AppointmentDaoImpl appointmentsList = new AppointmentDaoImpl();
-		ObservableList<Appointments> appointments;
+		ObservableList<Appointment> appointments;
 		/**
 		 * Try block to catch any SQL exceptions, otherwise get all appointments from the database
 		 */
 		try {
 			appointments = appointmentsList.getAll();
 		} catch (SQLException e) {
-			throw new RuntimeException(e);
+			throw new RuntimeException("Error getting all appointments", e);
 		}
-
 		appointmentIDColumn.setCellValueFactory(new PropertyValueFactory<>("appointmentID"));
 		titleColumn.setCellValueFactory(new PropertyValueFactory<>("title"));
 		descriptionColumn.setCellValueFactory(new PropertyValueFactory<>("description"));
@@ -109,6 +166,8 @@ public class MainController implements Initializable {
 		customerIDColumn.setCellValueFactory(new PropertyValueFactory<>("customerID"));
 		userIDColumn.setCellValueFactory(new PropertyValueFactory<>("userID"));
 		appointmentsTable.setItems(appointments);
+
+
 	}
 
 	/**
@@ -129,9 +188,23 @@ public class MainController implements Initializable {
 	 * Update appt btn.
 	 *
 	 * @param actionEvent the action event
+	 * @throws IOException the io exception
 	 */
-	public void updateApptBtn(ActionEvent actionEvent) {
-		FXMLLoader loader = new FXMLLoader(getClass().getResource("/view/UpdateApptScene.fxml"));
+	public void updateApptBtn(ActionEvent actionEvent) throws IOException {
+		selectedAppointment = appointmentsTable.getSelectionModel().getSelectedItem();
+		try {
+			if (selectedAppointment == null) {
+				throw new NullPointerException();
+			} else {
+				FXMLLoader loader = new FXMLLoader(getClass().getResource("/view/UpdateApptScene.fxml"));
+				Stage stage = (Stage) updateApptBtn.getScene().getWindow();
+				Scene scene = new Scene(loader.load());
+				stage.setTitle("Appointment Scheduler");
+				stage.setScene(scene);
+			}
+		} catch (NullPointerException e) {
+			ErrMsg.noApptorCustSelected("Please select an appointment to update", "No appointment selected");
+		}
 	}
 
 	/**
@@ -141,9 +214,17 @@ public class MainController implements Initializable {
 	 * @throws SQLException the sql exception
 	 */
 	public void deleteApptBtn(ActionEvent actionEvent) throws SQLException {
-		AppointmentDaoImpl appointmentDao = new AppointmentDaoImpl();
-		Appointments appointment = appointmentsTable.getSelectionModel().getSelectedItem();
-		appointmentDao.delete(appointment);
-		appointmentsTable.getItems().remove(appointment);
+		try {
+			if (appointmentsTable.getSelectionModel().getSelectedItem() != null) {
+				AppointmentDaoImpl appointmentDao = new AppointmentDaoImpl();
+				Appointment appointment = appointmentsTable.getSelectionModel().getSelectedItem();
+				appointmentDao.delete(appointment);
+				appointmentsTable.getItems().remove(appointment);
+			} else {
+				throw new NullPointerException();
+			}
+		} catch (NullPointerException e) {
+			ErrMsg.noApptorCustSelected("Please select an appointment to delete", "No appointment selected");
+		}
 	}
 }
